Update the unit test that checks environment variable expansion is supported
